refactor: relocate listen service
