+++ sidetrack/services/__init__.py
@@ -1,1 +1,14 @@
 """Service layer utilities."""
+
+from .listenbrainz import ListenBrainzClient, get_listenbrainz_client
+from .listens import ListenService, get_listen_service
+from .spotify import SpotifyClient, get_spotify_client
+
+__all__ = [
+    "ListenService",
+    "get_listen_service",
+    "SpotifyClient",
+    "get_spotify_client",
+    "ListenBrainzClient",
+    "get_listenbrainz_client",
+]
