+++ tests/services/test_listen_service_benchmark.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from sidetrack.api.services.listen_service import ListenService
+from sidetrack.services.listens import ListenService
 
 
 def _make_rows(n: int) -> list[dict]:
